move completion metadata to it's own file
